dutluk/dutluk_rs/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
fastapi==0.104.1
gensim==4.3.2
h11==0.14.0
idna==3.4
joblib==1.3.2
loguru==0.7.2
numpy==1.26.1
pandas==2.1.3
pinecone-client==2.2.4
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
